% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin}
\alias{bin}
\title{bin function}
\usage{
bin(n)
}
\arguments{
\item{n}{a non-negative integer}
}
\value{
the binary representation stored in a vector.
}
\description{
A function to convert decimal to binary
}
