% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_legend.R
\name{colour_legend}
\alias{colour_legend}
\title{colour_legend function}
\usage{
colour_legend(palette, suffix = "", dir = ".")
}
\arguments{
\item{palette}{X}

\item{suffix}{X}

\item{dir}{X}
}
\value{
...
}
\description{
A function to
}
