% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont_res.R
\name{cont_res}
\alias{cont_res}
\title{results concatenation}
\usage{
cont_res(pattern, res.fld, filter.EPIC = F)
}
\arguments{
\item{pattern}{file prefix of the results files}

\item{res.fld}{folder that contains the result files}

\item{filter.EPIC}{do filtering for EPIC chip data?}
}
\description{
This function allows you to concatenate the results generated from mcLME()
}
\examples{
cont_res('PLEvsCLE_','results')
}
\keyword{LME}
