% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taylor_correlation.R
\name{cor_taylor}
\alias{cor_taylor}
\title{cor_taylor function}
\usage{
cor_taylor(X)
}
\arguments{
\item{X}{a numeric matrix with number of rows bigger than the number of columns}
}
\value{
Taylor's correlation coefficient, a number between 0 and 1 expressing the amount of dependence between multiple variables.
}
\description{
A function to compute Taylor's correlation coefficient ;-)
}
