% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_diagnistics.R
\name{dplot}
\alias{dplot}
\title{dplot function}
\usage{
dplot(mod, ...)
}
\arguments{
\item{mod}{an object}

\item{...}{additional arguments}
}
\value{
method dplot
}
\description{
Generic function for model diagnostics.
}
