% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generic.R
\name{generic}
\alias{generic}
\title{generic function}
\usage{
generic(gen, methods = NULL, sp = 3, oname = "obj")
}
\arguments{
\item{gen}{character string giving the name of an S3 generic.}

\item{methods}{character vector: a list of methods for which to provide templates}

\item{sp}{the amont of space to put in between functions}

\item{oname}{name of the generic object}
}
\value{
roxygen text printed to the console.
}
\description{
A function to generate roxygen templates for generic funtions and associated methods.
}

