\name{getwikicoords}
\alias{getwikicoords}
\title{getwikicoords function...}
\usage{getwikicoords(place, county, rmslash=TRUE)}
\description{getwikicoords function}
\details{A function to return the lat/lon coordinates of towns in the UK from Wikipedia. Does not always work. Sometimes
the county has to be specified too.}
\value{The lat/lon coordinates from Wikipedia}
\arguments{\item{place}{character, ther name of the town}
\item{county}{character, the county it is in}
\item{rmslash}{remove slash from place name. Not normally used.}}
