% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garchForecast.R
\name{hCreate}
\alias{hCreate}
\title{hCreate function}
\usage{
hCreate(pars, y, T = length(y))
}
\arguments{
\item{pars}{parameters for the GARCH model, these would come from an MCMC run}

\item{y}{vector of log returns}

\item{T}{this is the length of y; allow this to be pre-computed}
}
\value{
vector of h's
}
\description{
A function used in the forecasting of GARCH(1,1) models
}
