\name{KFtemplates}
\alias{KFtemplates}
\title{KFtemplates function...}
\description{KFtemplates function}
\details{A function to print KFfit and KFparest templates to the console. See vignette("miscFuncs")
for more information}
\value{Tust prints to the console. This can be copied and pasted into a text editor for further manipulation.}
