% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garchForecast.R
\name{lr2fact}
\alias{lr2fact}
\title{lr2fact function}
\usage{
lr2fact(mod)
}
\arguments{
\item{mod}{the output of fcastGARCH}
}
\value{
the multiplicative factors.
}
\description{
Apply this to the output of fcastGARCH in order to undertake forecasting on the scale of the original series (i.e. not the log returns). Quantiles may
be computed across the MCMC iterations and then all one needs to do is to multiply the result by the last observed
value in the original series (again, not the log returns)
}
