% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcRLM.R
\name{mcRLM}
\alias{mcRLM}
\title{RLM in parallel}
\usage{
mcRLM(
  pDat,
  form,
  var.interest,
  chunk.list,
  chunk.prefix,
  output.prefix,
  num.chunks = 32,
  mc.core = 32,
  Feature_Name = NULL,
  Sample_Name = NULL
)
}
\arguments{
\item{pDat}{phenotype table (in the long format if it is longitudinal)}

\item{form}{model formula for rlm}

\item{var.interest}{coefficient to extract (append the non-reference level after in the case of a factor, eg. SexFemale)}

\item{chunk.list}{file names of your chunks}

\item{chunk.prefix}{folder that contains your chunks, make sure it ends with a '/'}

\item{output.prefix}{folder name and file name prefix of your result files}

\item{num.chunks}{number of chunks you have}

\item{mc.core}{cores that you want to use}

\item{Feature_Name}{name of the column that contains the feature names (in each chunk)}

\item{Sample_Name}{name of the column that contains the sample ID (in the phenotype table)}
}
\description{
This function allows you to run RLM in parallel for high dimensional data (methylation, SNP, expression...)
}
\details{
The features that you are working with will be called 'FeatureVal' within the function.
Make sure you use 'FeatureVal' in 'form' (model formula) and 'var.interest' (coefficient to extract, if applicable).
}
\examples{
chunk.list=paste0('chunk_',1:32,'.csv')

mcLM(pDat, form = 'Exprs ~ FeatureVal*Group',
var.interest = 'FeatureVal:GroupCLE', chunk.list=chunk.list,
chunk.prefix = 'processed_data/Beta_chunks/',
output.prefix='tmpres/PLECLEInt_',
num.chunks=32, mc.core=32, Feature_Name='cpg_names', Sample_Name='Sample_Name')
}
\keyword{RLM}
