% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{method}
\alias{method}
\title{method function}
\usage{
method(meth, gen, oname = "obj")
}
\arguments{
\item{meth}{character, the name of the method}

\item{gen}{character the associated generic method}

\item{oname}{name of object}
}
\value{
a roxygen template for the method.
}
\description{
A function to generate a roxygen template for a method of a generic S3 function. Normally, this would be called from
the function generic, see ?generic
}
