% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/22tables.R
\name{print22}
\alias{print22}
\title{print22 function}
\usage{
print22()
}
\value{
prints the names of the arguments of twotwofunction info to screen in their correct place in the 2 by 2 table
}
\description{
A function to print details of the 2 by 2 table for use with the function twotwoinfo.
}
\seealso{
\link{twotwoinfo}
}

