% Generated by roxygen2 (4.0.1): do not edit by hand
\name{roxbc}
\alias{roxbc}
\title{roxbc function}
\usage{
roxbc(name, checkflags = "--as-cran")
}
\arguments{
\item{name}{package name}

\item{checkflags}{string giving optional check flags to R CMD check, default is --as-cran}
}
\value{
builds and checks the package
}
\description{
A function to build and check packages where documentation has been compiled with roxygen. Probably only works in Linux.
}

