% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/roxbc.R
\name{roxbuild}
\alias{roxbuild}
\title{roxbuild function}
\usage{
roxbuild(name)
}
\arguments{
\item{name}{package name}
}
\value{
builds and checks the package
}
\description{
A function to build  packages where documentation has been compiled with roxygen. Probably only works in Linux.
}

