% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxtext.R
\name{roxtext}
\alias{roxtext}
\title{roxtext function}
\usage{
roxtext(fname)
}
\arguments{
\item{fname}{character string, the name of a function}
}
\value{
minimal roxygen template
}
\description{
A function to generate roxygen documentation templates for functions for example,
}
\details{
would generate a template for this function. Note that functions with default arguments that include quotes
will throw up an error at the moment, just delete these bits from the string, and if shold work.
}
