% Generated by roxygen2 (4.0.1): do not edit by hand
\name{timeop}
\alias{timeop}
\title{timeop function}
\usage{
timeop(expr)
}
\arguments{
\item{expr}{an expression to evaluate}
}
\value{
The time it took to evaluate the expression in seconds
}
\description{
A function to time an operation in R
}

