% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdc.R
\name{vdc}
\alias{vdc}
\title{vdc function}
\usage{
vdc(n)
}
\arguments{
\item{n}{the length of the sequence}
}
\value{
Van der Corput sequence of length n
}
\description{
A function to generate a Van der Corput sequence of numbers.
}

