% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_diagnistics.R
\name{vif}
\alias{vif}
\title{vif function}
\usage{
vif(mod)
}
\arguments{
\item{mod}{an object of class 'lm'}
}
\value{
...
}
\description{
Function to calculate the variance inflation factor for each variable in a linear regression model.
}
