% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garchForecast.R
\name{yhIterate}
\alias{yhIterate}
\title{yhIterate function}
\usage{
yhIterate(i, current, pars, eps, omega)
}
\arguments{
\item{i}{the index of the forward lags}

\item{current}{current matrix of (y,h)}

\item{pars}{parameters for the GARCH model, these would come from an MCMC run}

\item{eps}{matrix of Gaussian noise, dimension equal to number of MCMC iterations by the number of forecast lags}

\item{omega}{matrix of Inverse Gamma noise, dimension equal to number of MCMC iterations by the number of forecast lags}
}
\value{
two column matrix containing forecast y (1st column) and updated h (2nd column)
}
\description{
A function to perform forecasting of the series, used by fcastGARCH
}
