\name{generic}
\alias{generic}
\title{generic function...}
\usage{generic(gen, methods, sp=3, oname="obj")}
\description{generic function}
\details{A function to generate roxygen templates for generic funtions and associated methods.}
\value{roxygen text printed to the console.}
\arguments{\item{gen}{character string giving the name of an S3 generic.}
\item{methods}{character vector: a list of methods for which to provide templates}
\item{sp}{the amont of space to put in between functions}
\item{oname}{name of the generic object}}
