% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getwikicoords.R
\name{getwikicoords}
\alias{getwikicoords}
\title{getwikicoords function}
\usage{
getwikicoords(place, county = NULL, rmslash = TRUE)
}
\arguments{
\item{place}{character, ther name of the town}

\item{county}{character, the county it is in}

\item{rmslash}{remove slash from place name. Not normally used.}
}
\value{
The lat/lon coordinates from Wikipedia
}
\description{
A function to return the lat/lon coordinates of towns in the UK from Wikipedia. Does not always work. Sometimes
the county has to be specified too.
}

