% Generated by roxygen2 (4.0.1): do not edit by hand
\name{KFtemplates}
\alias{KFtemplates}
\title{KFtemplates function}
\usage{
KFtemplates()
}
\value{
Tust prints to the console. This can be copied and pasted into a text editor for further manipulation.
}
\description{
A function to print KFfit and KFparest templates to the console. See vignette("miscFuncs")
for more information
}

