% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latextable.R
\name{latextable}
\alias{latextable}
\title{latextable function}
\usage{
latextable(x, digits = 3, scientific = -3, colnames = NULL,
  rownames = NULL, caption = NULL, narep = " ", laststr = "", ...)
}
\arguments{
\item{x}{a matrix, or object that can be coerced to a matrix. x can include mixed character and numeric entries.}

\item{digits}{see help file for format}

\item{scientific}{see help file for format}

\item{colnames}{optional column names set to NULL (default) to automatically use column names of x. NOTE! if rownames is not NULL present, colnames must include an entry for the rownames i.e. it should be a vector of length the number of columns of x plus 1.}

\item{rownames}{optional row names set to NULL (default) to automatically use row names of x}

\item{caption}{optional caption, not normally used}

\item{narep}{string giving replacement for NA entries in the matrix}

\item{laststr}{string to write at end, eg note the double backslash!!}

\item{...}{additional arguments passed to format}
}
\value{
prints the LaTeX table to screen, so it can be copied into reports
}
\description{
A very useful function to create a LaTeX table from a matrix. Rounds numeric entries and also replaces
small numbers with standard index form equivalents.
}
\details{
To get a backslash to appear, use a double backslash

Just copy and paste the results into your LaTeX document.
}
\examples{
latextable(as.data.frame(matrix(1:4,2,2)))
}

