\name{method}
\alias{method}
\title{method function...}
\usage{method(meth, gen, oname="obj")}
\description{method function}
\details{A function to generate a roxygen template for a method of a generic S3 function. Normally, this would be called from
the function generic, see ?generic}
\value{a roxygen template for the method.}
\arguments{\item{meth}{character, the name of the method}
\item{gen}{character the associated generic method}
\item{oname}{name of object}}
