\name{roxbc}
\alias{roxbc}
\title{roxbc function}
\usage{
roxbc(name)
}
\arguments{
  \item{name}{package name}
}
\value{
builds and checks the package
}
\description{
A function to build and check packages where documentation
has been compiled with roxygen. Probably only works in
Linux. You will need to clean up old directorys
<package_name>.roxygen and <package_name>.Rcheck
}

