\name{roxbc}
\alias{roxbc}
\title{roxbc function...}
\usage{roxbc(name)}
\description{roxbc function}
\details{A function to build and check packages where documentation has been compiled with roxygen. Probably only works in Linux.
You will need to clean up old directorys <package_name>.roxygen and <package_name>.Rcheck}
\value{builds and checks the package}
\arguments{\item{name}{package name}}
