\name{roxtext}
\alias{roxtext}
\title{roxtext function}
\usage{
  roxtext(s)
}
\arguments{
  \item{s}{a string enclosed in quotes}
}
\value{
  minimal roxygen template
}
\description{
  A function to generate roxygen documentation templates
  for functions for example,
}
\details{
  would generate a template for this function. Note that
  functions with default arguments that include quotes will
  throw up an error at the moment, just delete these bits
  from the string, and if shold work.
}

